feat(front): make app file to use router and redux from synchronized data with frontend/ Additional info: This is a POC. What it does is by files: frontend/ OpenContentApp.jsx: it adds the redux store, by reference, to the data sent to apps Header.jsx: it adds a display of the redux state user.profileAvatarName that will be changed by app File. It also adds a display of a matching route that will be pushed by app File.
frontend_app_file/
package.json and yarn.lock: it adds the libraries redux and react-redux at the same version as the frontend/
index.js: it adds a <Provider> component that gets initialized with the store from the frontend
File.jsx: it adds the 2 HOC withRouter and connect(). It adds 2 buttons, one that will trigger an action to change the property profileAvatarName of state user in frontend's redux store. And one that will trigger a redirection to the route that will trigger an update of both frontend's Header.jsx and local File.jsx. It also displays the value frontend's redux state user.profileAvatarName to see it changing in app File.